VueRouter installed and set
